AEA-259: Added schema validation tests for sandbox mock requests. Started making the requests conform to the schemas.
