+++ specification/components/schemas/GroupIdentifier.yaml
@@ -31,8 +31,8 @@ properties:
     type: string
     description: Establishes the namespace for the value - that is, a URL that describes a set values that are unique.
     enum:
-      - "urn:oid:2.16.840.1.113883.2.1.3.2.4.18.8"
-    example: "urn:oid:2.16.840.1.113883.2.1.3.2.4.18.8"
+      - "https://fhir.nhs.uk/Id/prescription-short-form"
+    example: "https://fhir.nhs.uk/Id/prescription-short-form"
   value:
     type: string
     description: The portion of the identifier typically relevant to the user and which is unique within the context of the system.
